+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Support for modifying the request before sending it to the gateway.
+- Support for modifying the response before comparing it to the expected response.
 
 ## [0.2.0] - 2023-06-26
 ### Added